truncate links when displaying
Again, this is with the goal of avoiding the horizontal scrollbar. I'm
not sure if the extra complexity is worthwhile, though.

This took longer than I wanted because I started out with the idea of
stripping out utm_* tracking parameters because they seem to be the
major reason for long links. But it's worth a separate stage for the
simple thing when it's already adding so much complexity. Solve the
immediate problem first, try to avoid the tracking later.